fix(movie-catalog): code refactor

+++ movie-catalog/src/main/java/ru/walkername/movie_catalog/repositories/MoviesRepository.java
@@ -7,7 +7,7 @@ import ru.walkername.movie_catalog.models.Movie;
 import java.util.List;
 
 @Repository
-public interface MoviesRepository extends JpaRepository<Movie, Integer> {
+public interface MoviesRepository extends JpaRepository<Movie, Long> {
 
     List<Movie> findByTitleStartingWith(String title);
 }
